+++ src/flt-document-validator.ts
@@ -3,6 +3,7 @@ import {Diagnostic, DiagnosticCollection, TextDocument} from 'vscode';
 import {FtlFile} from './models/ftl-file';
 import {Validator} from './validators/validator';
 import {FtlRoot} from './models/ftl-root';
+import {DiagnosticBuilder} from './diagnostic-builder';
 
 export class FltDocumentValidator {
   constructor(private documentCache: DocumentCache,
@@ -11,7 +12,7 @@ export class FltDocumentValidator {
   }
 
   public validateFtlRoot(root: FtlRoot) {
-    this.validateFtlFiles(Array.from(root.files.values()));
+    this.validateFtlFiles(Array.from(root.xmlFiles.values()));
   }
 
   private validateFtlFiles(files: FtlFile[]) {
@@ -24,15 +25,20 @@ export class FltDocumentValidator {
 
   public validateFile(file: FtlFile) {
     const diagnostics: Diagnostic[] = [];
-    diagnostics.push(...file.diagnostics);
-    for (const validator of this.validators) {
-      validator.validateFile(file, diagnostics);
+    if (file.isReferenced) {
+      diagnostics.push(...file.diagnostics);
+      for (const validator of this.validators) {
+        validator.validateFile(file, diagnostics);
+      }
+    } else {
+      diagnostics.push(DiagnosticBuilder.fileNotReferenced(file.firstLineRange));
     }
+
     this.diagnosticCollection.set(file.uri, diagnostics);
   }
 
   public validateDocument(document: TextDocument, root: FtlRoot) {
-    const ftlFiles = root.files;
+    const ftlFiles = root.xmlFiles;
     const file = ftlFiles.get(document.uri.toString());
     if (file) {
       this.validateFile(file);